improvement(apollo-config): moved expanValue into separate file, and rename it to apolloStateToObject
